requirements.txt
pandas==1.3.3
numpy==1.20.3
Pillow==8.3.1
streamlit==1.2.0
spacy==2.3.5